requirements.txt
# Fire-ViT Requirements
# PyTorch and Vision
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# COCO Tools
pycocotools>=2.0.6

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0

# Utilities
einops>=0.6.0
scipy>=1.10.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0

# Optional: Deformable Attention
# mmcv-full>=1.7.0  # Uncomment if using MMDetection's deformable attention
